Accept Invitation change
