--- services/tenant-ui/frontend/src/views/connections/AcceptInvitation.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import AcceptInvite from '../../components/contacts/AcceptInvite.vue';
-</script>
-
-<template>
-  <AcceptInvite />
-</template>
